+++ sources/custom/ft_striteri.c
@@ -6,13 +6,13 @@
 /*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/05 15:35:24 by maolivei          #+#    #+#             */
-/*   Updated: 2022/04/08 17:16:00 by maolivei         ###   ########.fr       */
+/*   Updated: 2022/08/15 11:31:20 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_striteri(char *s, void (*f)(unsigned int, char*))
+void	ft_striteri(char *s, void (*f)(size_t, char*))
 {
 	size_t	i;
